refactor: remove unneeded deps
The palaver dep was only used for one line that could be replaced with
the Rust std lib. I also switched the hasher from the twox-hash library
to the default hasher. I'm aware that Rust makes no guarantees that the
internals of the hasher will stay the same between Rust versions, but
this is acceptable as different Rust versions will most likely produce
different libraries.